Remove AG0024UnitTests.cs and add TestCases (proof of concept)
