+++ src/Agoda.Analyzers.Test/AgodaCustom/TestCases/AG0024/NoWarning/FactoryStartNewWithLongRunningParameter.cs
@@ -1,0 +1,14 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+class FactoryStartNewWithLongRunningParameter
+{
+    public void TestMethod1()
+    {
+        Task.Factory.StartNew(MyMethod, CancellationToken.None, TaskCreationOptions.LongRunning, TaskScheduler.Default);
+    }
+
+    public void MyMethod()
+    {
+    }
+}
