+++ src/Agoda.Analyzers.Test/AgodaCustom/TestCases/AG0024/Warning/FactoryStartNewWithMethodParameter.cs
@@ -1,0 +1,15 @@
+﻿/*9,9*/
+using System.Threading;
+using System.Threading.Tasks;
+
+class FactoryStartNewWithMethodParameter
+{
+    public void TestMethod1()
+    {
+        Task.Factory.StartNew(MyMethod);
+    }
+
+    public void MyMethod()
+    {
+    }
+}
