+++ src/Agoda.Analyzers.Test/AgodaCustom/TestCases/AG0024/Warning/FactoryStartNewWithMultipleParameters.cs
@@ -1,0 +1,15 @@
+﻿/*9,9*/
+using System.Threading;
+using System.Threading.Tasks;
+
+class FactoryStartNewWithMultipleParameters
+{
+    public void TestMethod1()
+    {
+        Task.Factory.StartNew(MyMethod, CancellationToken.None, TaskCreationOptions.DenyChildAttach, TaskScheduler.Default);
+    }
+
+    public void MyMethod()
+    {
+    }
+}
